TDLS-1: Restructure (since it is a small app)
